back-end/requirements.txt
django>=5.1,<5.2
whitenoise>=6.8,<6.9
djangorestframework>=3.15,<3.16
djangorestframework-simplejwt>=5.4,<5.5
psycopg2-binary>=2.9,<2.10
gunicorn>=23.0,<24.0
django-filter>=24.3,<24.4
django-cors-headers>=4.6,<4.7
faker>=33.0,<34.0
celery>=5.3
redis>=5.0
httpx>=0.27
django-celery-results>=2.5
django-celery-beat>=2.6
